Changed to use mocked repositories in web applications instead of mocked repositories
