+++ src/Chirp.Infrastructure/Contexts/ChirbDbContext.cs
@@ -14,7 +14,10 @@ public class ChirpDbContext : DbContext
 
     public ChirpDbContext(DbContextOptions<ChirpDbContext> options)
         : base(options)
-    { }
+    {
+        Database.Migrate();
+        DbInitializer.SeedDatabase(this);
+    }
     
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
