+++ test/Chirp.Tests.Core/MockRepositoryFactory.cs
@@ -26,53 +26,26 @@ public struct MockChirpRepositories
 
 public static class MockRepositoryFactory
 {
-    public static MockChirpRepositories GetMockCheepRepository(bool withErrorProvocation = false)
+    public static MockChirpRepositories GetMockCheepRepositories(bool withErrorProvocation = false)
     {
-        var mockAuthorsDbSet = new Mock<DbSet<Author>>();
-        var mockCheepsDbSet = new Mock<DbSet<Cheep>>();
-        var mockLikesDbSet = new Mock<DbSet<Like>>();
-        
-        DataGenerator.AuthorCheepsData data = DataGenerator.GenerateAuthorsAndCheeps();
-        
-        // If mock db sets are not set up, it will throw an exception which will be caught by FetchWithErrorHandling
-        if (!withErrorProvocation) {
-            mockAuthorsDbSet.As<IQueryable<Author>>().Setup(m => m.Provider).Returns(data.Authors.AsQueryable().Provider);
-            mockAuthorsDbSet.As<IQueryable<Author>>().Setup(m => m.Expression).Returns(data.Authors.AsQueryable().Expression);
-            mockAuthorsDbSet.As<IQueryable<Author>>().Setup(m => m.ElementType).Returns(data.Authors.AsQueryable().ElementType);
-            mockAuthorsDbSet.As<IQueryable<Author>>().Setup(m => m.GetEnumerator()).Returns(data.Authors.AsQueryable().GetEnumerator());
-            
-            mockCheepsDbSet.As<IQueryable<Cheep>>().Setup(m => m.Provider).Returns(data.Cheeps.AsQueryable().Provider);
-            mockCheepsDbSet.As<IQueryable<Cheep>>().Setup(m => m.Expression).Returns(data.Cheeps.AsQueryable().Expression);
-            mockCheepsDbSet.As<IQueryable<Cheep>>().Setup(m => m.ElementType).Returns(data.Cheeps.AsQueryable().ElementType);
-            mockCheepsDbSet.As<IQueryable<Cheep>>().Setup(m => m.GetEnumerator()).Returns(data.Cheeps.AsQueryable().GetEnumerator());
+        var mockChirpDbContext = DataGenerator.GenerateMockChirpDbContext(withErrorProvocation);
 
-            mockLikesDbSet.As<IQueryable<Like>>().Setup(m => m.Provider).Returns(data.Likes.AsQueryable().Provider);
-            mockLikesDbSet.As<IQueryable<Like>>().Setup(m => m.Expression).Returns(data.Likes.AsQueryable().Expression);
-            mockLikesDbSet.As<IQueryable<Like>>().Setup(m => m.ElementType).Returns(data.Likes.AsQueryable().ElementType);
-            mockLikesDbSet.As<IQueryable<Like>>().Setup(m => m.GetEnumerator()).Returns(data.Likes.AsQueryable().GetEnumerator());
-        }
-        
-        var mockChirpDbContext = new Mock<ChirpDbContext>();
-        mockChirpDbContext.Setup(m => m.Authors).Returns(mockAuthorsDbSet.Object);
-        mockChirpDbContext.Setup(m => m.Cheeps).Returns(mockCheepsDbSet.Object);
-        mockChirpDbContext.Setup(m => m.Likes).Returns(mockLikesDbSet.Object);
-
-        IAuthorRepository authorRepository = new AuthorRepository(mockChirpDbContext.Object);
-        ICheepRepository cheepRepository = new CheepRepository(mockChirpDbContext.Object, authorRepository);
-        ILikeRepository likeRepository = new LikeRepository(mockChirpDbContext.Object);
+        IAuthorRepository authorRepository = new AuthorRepository(mockChirpDbContext.MockChirpDbContext.Object);
+        ICheepRepository cheepRepository = new CheepRepository(mockChirpDbContext.MockChirpDbContext.Object, authorRepository);
+        ILikeRepository likeRepository = new LikeRepository(mockChirpDbContext.MockChirpDbContext.Object);
         
         return new MockChirpRepositories
         {
             CheepRepository = cheepRepository,
             AuthorRepository = authorRepository,
             LikeRepository = likeRepository,
-            TestAuthors = data.Authors,
-            TestCheeps = data.Cheeps,
-            TestLikes = data.Likes,
-            MockChirpDbContext = mockChirpDbContext,
-            MockAuthorsDbSet = mockAuthorsDbSet,
-            MockCheepsDbSet = mockCheepsDbSet,
-            MockLikesDbSet = mockLikesDbSet,
+            TestAuthors = mockChirpDbContext.Authors,
+            TestCheeps = mockChirpDbContext.Cheeps,
+            TestLikes = mockChirpDbContext.Likes,
+            MockChirpDbContext = mockChirpDbContext.MockChirpDbContext,
+            MockAuthorsDbSet = mockChirpDbContext.MockDbAuthorsSet,
+            MockCheepsDbSet = mockChirpDbContext.MockDbCheepsSet,
+            MockLikesDbSet = mockChirpDbContext.MockDbLikesSet,
         };
     }
 }
